Start using DSA where we can, and fix an issue with applying filtering at runtime.
So, where to start here? PLG operates on the current object, unlike OpenGL which has a global state. Unfortunately, this idiot didn't account for that very well. When we were setting up the current draw pass in Ape, we bound the texture to the desired texture unit, but immediately after that, PLG changes the current unit back to 0, and we then try changing the texture filter... UH OH, can you guess what went wrong? Yes! The texture is then bound again, but this time to the wrong texture unit and the original texture wasn't bound again.

DSA avoids these idiotic issues since we don't need to worry about OpenGL's internal state so much, but obviously depends on OpenGL 4.5+ (which these days is fairly well supported).

On another note this also simplifies the NOMIPS flag; rather than the client applying this to the texture to indicate if they want mipmaps or not, it's instead used on our side to indicate if the texture has mipmaps or not - if the caller doesn't want mipmaps this is now done by specifying the filtering type for the texture on creation instead prior to upload.

Signed-off-by: Mark Sowden <hogsy@oldtimes-software.com>